Refactor authorization to use Azure SDK
